# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CarrierCatalogOut": {
        "description": "Every carrier this BUILD supports, live or not - so an operator can see what is\navailable to switch on, not merely what is already on.",
        "properties": {
          "capabilities": {
            "additionalProperties": true,
            "default": {},
            "title": "Capabilities",
            "type": "object"
          },
          "enabled_flag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled Flag"
          },
          "live": {
            "title": "Live",
            "type": "boolean"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary": {
            "default": false,
            "title": "Primary",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "default": "",
            "title": "State",
            "type": "string"
          },
          "supports_messaging": {
            "default": true,
            "title": "Supports Messaging",
            "type": "boolean"
          },
          "supports_numbers": {
            "default": false,
            "title": "Supports Numbers",
            "type": "boolean"
          },
          "supports_voice": {
            "default": false,
            "title": "Supports Voice",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "live",
          "reason",
          "missing",
          "enabled_flag"
        ],
        "title": "CarrierCatalogOut",
        "type": "object"
      },
      "CarrierStatusOut": {
        "properties": {
          "capabilities": {
            "additionalProperties": true,
            "title": "Capabilities",
            "type": "object"
          },
          "consecutive_failures": {
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary": {
            "title": "Primary",
            "type": "boolean"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "name",
          "primary",
          "state",
          "consecutive_failures",
          "capabilities"
        ],
        "title": "CarrierStatusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyIn": {
        "properties": {
          "allow_cross_carrier_failover": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Cross Carrier Failover"
          },
          "allow_intra_carrier_failover": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Intra Carrier Failover"
          },
          "pinned_carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Carrier"
          },
          "preference": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference"
          }
        },
        "title": "PolicyIn",
        "type": "object"
      },
      "PolicyOut": {
        "properties": {
          "allow_cross_carrier_failover": {
            "title": "Allow Cross Carrier Failover",
            "type": "boolean"
          },
          "allow_intra_carrier_failover": {
            "title": "Allow Intra Carrier Failover",
            "type": "boolean"
          },
          "pinned_carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Carrier"
          },
          "preference": {
            "items": {
              "type": "string"
            },
            "title": "Preference",
            "type": "array"
          }
        },
        "required": [
          "preference",
          "allow_intra_carrier_failover",
          "allow_cross_carrier_failover",
          "pinned_carrier"
        ],
        "title": "PolicyOut",
        "type": "object"
      },
      "ProbeOut": {
        "properties": {
          "checked": {
            "title": "Checked",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "ok",
          "detail",
          "checked"
        ],
        "title": "ProbeOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CSaaS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/routing/carriers": {
      "get": {
        "description": "Which carriers this deployment can actually use, and how each one is behaving.\n\nNever includes a credential - only whether one is present and working.",
        "operationId": "list_carriers_api_v1_routing_carriers_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CarrierStatusOut"
                  },
                  "title": "Response List Carriers Api V1 Routing Carriers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Carriers",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/carriers/{name}/probe": {
      "post": {
        "description": "Ask the carrier whether these credentials work. Operator-triggered ONLY - see\nproviders/probes.py for why this never runs on boot.",
        "operationId": "probe_carrier_api_v1_routing_carriers__name__probe_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Probe Carrier",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/catalog": {
      "get": {
        "description": "The bring-your-own-key view: every supported carrier, whether it is live, and if\nnot, the exact variables that would make it live.",
        "operationId": "carrier_catalog_api_v1_routing_catalog_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CarrierCatalogOut"
                  },
                  "title": "Response Carrier Catalog Api V1 Routing Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Carrier Catalog",
        "tags": [
          "routing"
        ]
      }
    },
    "/api/v1/routing/policy": {
      "get": {
        "operationId": "get_policy_api_v1_routing_policy_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Policy",
        "tags": [
          "routing"
        ]
      },
      "patch": {
        "operationId": "update_policy_api_v1_routing_policy_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Policy",
        "tags": [
          "routing"
        ]
      }
    },
...
  }
}